# docs/swagger/swagger.yaml
basePath: /
definitions:
...
  handler.MessageData:
    properties:
      message:
        example: success
        type: string
    type: object
...
  handler.RolePageData:
    properties:
      list:
        items:
          $ref: '#/definitions/service.RoleItem'
        type: array
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 1
        type: integer
    type: object
...
  response.Body:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
...
  service.PolicyGrantRequest:
    properties:
      permission_id:
        type: integer
      role_id:
        type: integer
    required:
    - permission_id
    - role_id
    type: object
  service.PolicyItemResponse:
    properties:
      action:
        type: string
      permission_id:
        type: integer
      permission_name:
        type: string
      resource:
        type: string
      role_code:
        type: string
      role_id:
        type: integer
      role_name:
        type: string
    type: object
  service.RoleCreateRequest:
    properties:
      code:
        maxLength: 64
        type: string
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 64
        type: string
      status:
        type: integer
    required:
    - code
    - name
    type: object
  service.RoleItem:
    properties:
      code:
        type: string
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      status:
        type: integer
    type: object
  service.RoleUpdateRequest:
    properties:
      code:
        maxLength: 64
        type: string
      description:
        maxLength: 255
        type: string
      name:
        maxLength: 64
        type: string
      status:
        type: integer
    type: object
...
info:
  contact: {}
  description: |-
    Permission management system built with Gin, MySQL, Redis, Casbin, Cobra and slog.
    Use the login endpoint first, then put `Bearer <token>` into the Authorization header.
  title: Permission System API
  version: "1.0"
paths:
...
  /api/v1/policies:
    delete:
      consumes:
      - application/json
      description: Remove one permission from one role.
      parameters:
      - description: Revoke policy request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.PolicyGrantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/handler.MessageData'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: resource not found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Revoke policy
      tags:
      - Policy
    get:
      description: List current role-permission policy bindings.
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/service.PolicyItemResponse'
                  type: array
              type: object
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: List policies
      tags:
      - Policy
    post:
      consumes:
      - application/json
      description: Bind one permission to one role.
      parameters:
      - description: Grant policy request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.PolicyGrantRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/handler.MessageData'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: resource not found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Grant policy
      tags:
      - Policy
  /api/v1/roles:
    get:
      description: List roles with optional keyword and pagination filters.
      parameters:
      - description: Search keyword
        in: query
        name: keyword
        type: string
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/handler.RolePageData'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: List roles
      tags:
      - Role
    post:
      consumes:
      - application/json
      description: Create a new role.
      parameters:
      - description: Create role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.RoleCreateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: created
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/service.RoleItem'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Create role
      tags:
      - Role
  /api/v1/roles/{id}:
    delete:
      description: Delete a role by ID.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/handler.MessageData'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: role not found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Delete role
      tags:
      - Role
    get:
      description: Get role detail by ID.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/service.RoleItem'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: role not found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Get role detail
      tags:
      - Role
    put:
      consumes:
      - application/json
      description: Update role name, code, description or status.
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update role request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/service.RoleUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: success
          schema:
            allOf:
            - $ref: '#/definitions/response.Body'
            - properties:
                data:
                  $ref: '#/definitions/service.RoleItem'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/response.Body'
        "401":
          description: unauthorized
          schema:
            $ref: '#/definitions/response.Body'
        "403":
          description: forbidden
          schema:
            $ref: '#/definitions/response.Body'
        "404":
          description: role not found
          schema:
            $ref: '#/definitions/response.Body'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/response.Body'
      security:
      - BearerAuth: []
      summary: Update role
      tags:
      - Role
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Enter the JWT token with the `Bearer ` prefix, for example: Bearer
      eyJhbGciOi...'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"